pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gozumol"
version = "0.1.0"
description = "AI-powered visual navigation assistant for visually impaired users"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Gozumol Team"}
]
keywords = ["vision", "accessibility", "navigation", "multimodal", "AI", "visually-impaired"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "accelerate>=0.25.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
phi4 = ["flash-attn>=2.5.0"]
qwen = ["qwen-vl-utils"]
camera = ["opencv-python>=4.8.0"]
tts = ["pyttsx3>=2.90", "gTTS>=2.3.0"]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/eminaruk/gozumol"
Documentation = "https://github.com/eminaruk/gozumol#readme"
Repository = "https://github.com/eminaruk/gozumol"
Issues = "https://github.com/eminaruk/gozumol/issues"

[project.scripts]
gozumol = "gozumol.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gozumol*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
